Bug Correction about pagination

+++ android/data/src/main/java/com/backbase/data/api/responses/MovieDetailsResult.kt
@@ -2,12 +2,13 @@ package com.backbase.data.api.responses
 
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
+import java.math.BigInteger
 
 data class MovieDetailsResult(
     @Expose @SerializedName("adult") val adult: Boolean?,
     @Expose @SerializedName("backdrop_path") val backdrop_path: String?,
     @Expose @SerializedName("belongs_to_collection") val belongs_to_collection: BelongsToCollection?,
-    @Expose @SerializedName("budget") val budget: Int?,
+    @Expose @SerializedName("budget") val budget: BigInteger?,
     @Expose @SerializedName("genres") val genres: List<GenreApi>?,
     @Expose @SerializedName("homepage") val homepage: String?,
     @Expose @SerializedName("id") val id: Int?,
@@ -20,7 +21,7 @@ data class MovieDetailsResult(
     @Expose @SerializedName("production_companies") val production_companies: List<ProductionCompanyApi>?,
     @Expose @SerializedName("production_countries") val production_countries: List<ProductionCountryApi>?,
     @Expose @SerializedName("release_date") val release_date: String?,
-    @Expose @SerializedName("revenue") val revenue: Int?,
+    @Expose @SerializedName("revenue") val revenue: BigInteger?,
     @Expose @SerializedName("runtime") val runtime: Int?,
     @Expose @SerializedName("spoken_languages") val spoken_languages: List<SpokenLanguageApi>?,
     @Expose @SerializedName("status") val status: String?,
@@ -28,7 +29,7 @@ data class MovieDetailsResult(
     @Expose @SerializedName("title") val title: String?,
     @Expose @SerializedName("video") val video: Boolean?,
     @Expose @SerializedName("vote_average") val vote_average: Double?,
-    @Expose @SerializedName("vote_count") val vote_count: Int?
+    @Expose @SerializedName("vote_count") val vote_count: BigInteger?
 )
 
 data class BelongsToCollection(
